visionIQ/python-workers-visionIQ/requirements-shared.txt
redis>=5.0.0
asyncpg>=0.29.0
qdrant-client>=1.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
